+++ server/impl/src/main/java/com/walmartlabs/concord/server/process/queue/ProcessQueueWatchdog.java
@@ -47,7 +47,6 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.UUID;
-import java.util.concurrent.TimeUnit;
 
 import static com.walmartlabs.concord.db.PgUtils.interval;
 import static com.walmartlabs.concord.server.jooq.tables.ProcessQueue.PROCESS_QUEUE;
@@ -132,7 +131,7 @@ public class ProcessQueueWatchdog implements ScheduledTask {
 
     @Override
     public long getIntervalInSec() {
-        return TimeUnit.SECONDS.toSeconds(2);
+        return cfg.getPeriod();
     }
 
     @Override